refactor: SettingsScreen を各カテゴリ Screen に分割
各設定カテゴリを独立した Screen composable として切り出し、
SettingsScreen / SettingsContent はナビゲーションに徹する構成に変更。

- AccountScreen: パスワード変更・パスキー管理・ログイン履歴
- CreditsScreen: OSS ライセンス・データクレジット
- UserManagementScreen: ユーザー名管理（admin）
- PetScreen: ペット名・ごはん設定・リマインダー（admin）
- GarbageScreen: ゴミ出しスケジュール・通知（admin）
- QuestScreen: Quest Webhook 設定（admin）
- MoneyScreen: お金・入金 Webhook 設定（admin）
- CacheScreen: サーバーキャッシュ（admin）

各 Screen が自分の ViewModel を koinViewModel() で直接注入するため、
SettingsContent のシグネチャが 70+ パラメータ → 4 パラメータに削減。
SettingsModule の admin ViewModel 登録を factory → viewModel に変更。
