+++ feature/settings/src/commonMain/kotlin/feature/settings/AccountScreen.kt
@@ -1,0 +1,266 @@
+package feature.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Fingerprint
+import androidx.compose.material.icons.filled.Lock
+import androidx.compose.material.icons.filled.Visibility
+import androidx.compose.material.icons.filled.VisibilityOff
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import org.koin.compose.viewmodel.koinViewModel
+
+@Composable
+internal fun AccountScreen(modifier: Modifier = Modifier) {
+    val passwordVm: PasswordChangeViewModel = koinViewModel()
+    val passkeyVm: PasskeyManagementViewModel = koinViewModel()
+    val loginHistoryVm: LoginHistoryViewModel = koinViewModel()
+    val s = passwordVm.uiState
+    val pk = passkeyVm.uiState
+    val lh = loginHistoryVm.uiState
+
+    PasswordChangeCard(
+        currentPassword = s.currentPassword,
+        newPassword = s.newPassword,
+        confirmPassword = s.confirmPassword,
+        isLoading = s.isLoading,
+        errorMessage = s.errorMessage,
+        successMessage = s.successMessage,
+        onCurrentPasswordChanged = passwordVm::onCurrentPasswordChanged,
+        onNewPasswordChanged = passwordVm::onNewPasswordChanged,
+        onConfirmPasswordChanged = passwordVm::onConfirmPasswordChanged,
+        onChangePassword = passwordVm::onChangePassword,
+        modifier = modifier,
+    )
+
+    if (pk.isAvailable) {
+        PasskeyManagementCard(
+            credentialCount = pk.credentialCount,
+            isRegistering = pk.isRegistering,
+            errorMessage = pk.errorMessage,
+            successMessage = pk.successMessage,
+            onRegisterPasskey = passkeyVm::onRegisterPasskey,
+            modifier = modifier,
+        )
+    }
+
+    LoginHistoryCardContent(
+        isLoading = lh.isLoading,
+        loadError = lh.loadError,
+        loadErrorMessage = lh.loadErrorMessage,
+        events = lh.events,
+        onRetry = loginHistoryVm::loadHistory,
+        modifier = modifier,
+    )
+}
+
+@Composable
+private fun PasswordChangeCard(
+    currentPassword: String,
+    newPassword: String,
+    confirmPassword: String,
+    isLoading: Boolean,
+    errorMessage: String?,
+    successMessage: String?,
+    onCurrentPasswordChanged: (String) -> Unit,
+    onNewPasswordChanged: (String) -> Unit,
+    onConfirmPasswordChanged: (String) -> Unit,
+    onChangePassword: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    var currentPasswordVisible by remember { mutableStateOf(false) }
+    var newPasswordVisible by remember { mutableStateOf(false) }
+    var confirmPasswordVisible by remember { mutableStateOf(false) }
+
+    Card(
+        modifier = modifier,
+        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant),
+    ) {
+        Column(
+            modifier = Modifier.padding(24.dp).fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
+            ) {
+                Icon(imageVector = Icons.Default.Lock, contentDescription = null, tint = MaterialTheme.colorScheme.primary)
+                Text(text = "パスワード変更", style = MaterialTheme.typography.titleSmall, color = MaterialTheme.colorScheme.onSurface)
+            }
+
+            OutlinedTextField(
+                value = currentPassword,
+                onValueChange = onCurrentPasswordChanged,
+                label = { Text("現在のパスワード") },
+                leadingIcon = { Icon(Icons.Default.Lock, contentDescription = null) },
+                trailingIcon = {
+                    IconButton(onClick = { currentPasswordVisible = !currentPasswordVisible }) {
+                        Icon(
+                            if (currentPasswordVisible) Icons.Default.VisibilityOff else Icons.Default.Visibility,
+                            contentDescription = if (currentPasswordVisible) "パスワードを隠す" else "パスワードを表示",
+                        )
+                    }
+                },
+                singleLine = true,
+                visualTransformation = if (currentPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password, imeAction = ImeAction.Next),
+                modifier = Modifier.fillMaxWidth(),
+                enabled = !isLoading,
+            )
+
+            OutlinedTextField(
+                value = newPassword,
+                onValueChange = onNewPasswordChanged,
+                label = { Text("新しいパスワード") },
+                leadingIcon = { Icon(Icons.Default.Lock, contentDescription = null) },
+                trailingIcon = {
+                    IconButton(onClick = { newPasswordVisible = !newPasswordVisible }) {
+                        Icon(
+                            if (newPasswordVisible) Icons.Default.VisibilityOff else Icons.Default.Visibility,
+                            contentDescription = if (newPasswordVisible) "パスワードを隠す" else "パスワードを表示",
+                        )
+                    }
+                },
+                singleLine = true,
+                visualTransformation = if (newPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password, imeAction = ImeAction.Next),
+                modifier = Modifier.fillMaxWidth(),
+                enabled = !isLoading,
+            )
+
+            OutlinedTextField(
+                value = confirmPassword,
+                onValueChange = onConfirmPasswordChanged,
+                label = { Text("新しいパスワード（確認）") },
+                leadingIcon = { Icon(Icons.Default.Lock, contentDescription = null) },
+                trailingIcon = {
+                    IconButton(onClick = { confirmPasswordVisible = !confirmPasswordVisible }) {
+                        Icon(
+                            if (confirmPasswordVisible) Icons.Default.VisibilityOff else Icons.Default.Visibility,
+                            contentDescription = if (confirmPasswordVisible) "パスワードを隠す" else "パスワードを表示",
+                        )
+                    }
+                },
+                singleLine = true,
+                visualTransformation = if (confirmPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password, imeAction = ImeAction.Done),
+                modifier = Modifier.fillMaxWidth(),
+                enabled = !isLoading,
+            )
+
+            if (errorMessage != null) {
+                Text(text = errorMessage, color = MaterialTheme.colorScheme.error, style = MaterialTheme.typography.bodySmall)
+            }
+
+            if (successMessage != null) {
+                Text(text = successMessage, color = MaterialTheme.colorScheme.primary, style = MaterialTheme.typography.bodySmall)
+            }
+
+            Button(
+                onClick = onChangePassword,
+                modifier = Modifier.height(48.dp),
+                enabled = !isLoading,
+            ) {
+                if (isLoading) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(20.dp),
+                        strokeWidth = 2.dp,
+                        color = MaterialTheme.colorScheme.onPrimary,
+                    )
+                } else {
+                    Text("変更する")
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun PasskeyManagementCard(
+    credentialCount: Int,
+    isRegistering: Boolean,
+    errorMessage: String?,
+    successMessage: String?,
+    onRegisterPasskey: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Card(
+        modifier = modifier,
+        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant),
+    ) {
+        Column(
+            modifier = Modifier.padding(24.dp).fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
+            ) {
+                Icon(imageVector = Icons.Default.Fingerprint, contentDescription = null, tint = MaterialTheme.colorScheme.primary)
+                Text(text = "パスキー管理", style = MaterialTheme.typography.titleSmall, color = MaterialTheme.colorScheme.onSurface)
+            }
+
+            Text(
+                text = "登録済み: $credentialCount 件",
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+            )
+
+            Text(
+                text = "別の端末やブラウザからログインするには、パスキーを追加してください。",
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+            )
+
+            if (errorMessage != null) {
+                Text(text = errorMessage, color = MaterialTheme.colorScheme.error, style = MaterialTheme.typography.bodySmall)
+            }
+
+            if (successMessage != null) {
+                Text(text = successMessage, color = MaterialTheme.colorScheme.primary, style = MaterialTheme.typography.bodySmall)
+            }
+
+            Button(
+                onClick = onRegisterPasskey,
+                modifier = Modifier.height(48.dp),
+                enabled = !isRegistering,
+            ) {
+                if (isRegistering) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(20.dp),
+                        strokeWidth = 2.dp,
+                        color = MaterialTheme.colorScheme.onPrimary,
+                    )
+                } else {
+                    Text("パスキーを追加")
+                }
+            }
+        }
+    }
+}
