+++ feature/settings/src/commonMain/kotlin/feature/settings/CacheScreen.kt
@@ -1,0 +1,80 @@
+package feature.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import org.koin.compose.viewmodel.koinViewModel
+
+@Composable
+internal fun CacheScreen(modifier: Modifier = Modifier) {
+    val vm: CacheRefreshViewModel = koinViewModel()
+    CacheRefreshCard(
+        isClearing = vm.uiState.isClearing,
+        message = vm.uiState.message,
+        onClearCache = vm::onClearCache,
+        modifier = modifier,
+    )
+}
+
+@Composable
+private fun CacheRefreshCard(
+    isClearing: Boolean,
+    message: String?,
+    onClearCache: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Card(
+        modifier = modifier,
+        colors =
+            CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.surfaceVariant,
+            ),
+    ) {
+        Column(
+            modifier = Modifier.padding(24.dp).fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+        ) {
+            Text(
+                text = "データの不整合が疑われる場合に、サーバー側のキャッシュを手動でクリアします。",
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+            )
+
+            if (message != null) {
+                Text(
+                    text = message,
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.primary,
+                )
+            }
+
+            Button(
+                onClick = onClearCache,
+                modifier = Modifier.height(48.dp),
+                enabled = !isClearing,
+            ) {
+                if (isClearing) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(20.dp),
+                        strokeWidth = 2.dp,
+                        color = MaterialTheme.colorScheme.onPrimary,
+                    )
+                } else {
+                    Text("キャッシュをクリア")
+                }
+            }
+        }
+    }
+}
